Added Friends to Shopping app.
Added a few new activities. A new Hub activity which will hold the main
interface of the app. Lists a few fake items and has a few fake settings. Also
added a list view of friends, and an activity to add to the friend list during
the session. Added new Friend class to eventually be how we interact with
friends. Also changed some styles to make it compatible with non-android 5
phones. This work was done for M5.